+++ containeranalysis/CHANGES.md
@@ -1,5 +1,12 @@
 # Changes
 
+## [0.12.2](https://github.com/googleapis/google-cloud-go/compare/containeranalysis/v0.12.1...containeranalysis/v0.12.2) (2024-08-08)
+
+
+### Bug Fixes
+
+* **containeranalysis:** Update google.golang.org/api to v0.191.0 ([5b32644](https://github.com/googleapis/google-cloud-go/commit/5b32644eb82eb6bd6021f80b4fad471c60fb9d73))
+
 ## [0.12.1](https://github.com/googleapis/google-cloud-go/compare/containeranalysis/v0.12.0...containeranalysis/v0.12.1) (2024-07-24)